pyproject.toml
[tool.poetry]
name = "go-champs-game-events"
version = "0.1.0"
description = "Project designed to manage game events"
authors = ["Eduardo Barreto <ecoutobarreto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
python-dotenv = "^1.0.1"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
